General tidy of code
