+++ src/recipes/tables/index.tsx
@@ -4,34 +4,19 @@ import {
   // DynamicTableProps,
   // DynamicTableState,
   InMemoryTableProps,
-  CreateTableHeader,
-  CreateTableRow,
-  CreateTableRows,
   InMemoryPaginatedTableProps,
-  CreatePageButtons,
-  CreateTableHeaderCell,
 } from "./types";
 
-import {
-  Table,
-  TableBody,
-  TableDataCell,
-  TableHead,
-  TableHeadCell,
-  TableRow,
-} from "../../html/Table";
+import { Table } from "../../html/Table";
 import { Stack } from "../../primitives/Stack";
-import { Inline } from "../../primitives/Inline";
-import { Text } from "../../primitives/Text";
-import { Button } from "../Button";
 import {
+  createTableHeader,
+  createTableBody,
   getHeadingKeys,
   getInitialSortState,
-  getNextSortState,
   getPage,
-  pageDown,
-  pageUp,
   sortRows,
+  createPaginationControls,
 } from "./utils";
 
 // TODO Probably want to define a type for table context here
@@ -52,105 +37,6 @@ export const TableContext = React.createContext({});
 
 //   return <TableContext.Provider value={context}></TableContext.Provider>;
 // };
-
-const createTableHeaderCell: CreateTableHeaderCell = ({
-  headingKey,
-  headerConfig,
-  sortState,
-  setSortState,
-}) => {
-  const { label, sortable } = headerConfig;
-  const { sortDirection } = sortState;
-
-  if (sortable) {
-    const sortButtonLabel = sortDirection === "ASCENDING" ? "A" : "D";
-    const nextSortState = getNextSortState({
-      sortState,
-      nextSortAttribute: headingKey,
-    });
-
-    return (
-      <TableHeadCell key={headingKey}>
-        <Inline>
-          <Text content={label}></Text>
-          <Button
-            label={sortButtonLabel}
-            onPress={() => setSortState(nextSortState)}
-          ></Button>
-        </Inline>
-      </TableHeadCell>
-    );
-  }
-  return (
-    <TableHeadCell>
-      <Text content={label}></Text>
-    </TableHeadCell>
-  );
-};
-
-const createTableHeader: CreateTableHeader<Object> = ({
-  tableHeaderConfig,
-  sortState,
-  setSortState,
-}) => {
-  const headerCells = [];
-  for (const [key, value] of Object.entries(tableHeaderConfig)) {
-    headerCells.push(
-      createTableHeaderCell({
-        headingKey: key,
-        headerConfig: value,
-        sortState,
-        setSortState,
-      })
-    );
-  }
-  return headerCells;
-};
-
-const createTableRow: CreateTableRow<Object> = ({
-  tableRowData,
-  headingKeys,
-  rowNumber,
-}) => {
-  return headingKeys.map((headingKey, index) => (
-    <TableDataCell key={`${rowNumber}_${index}`}>
-      {tableRowData[headingKey]}
-    </TableDataCell>
-  ));
-};
-
-const createTableRows: CreateTableRows<Object> = ({
-  tableData,
-  headingKeys,
-}) => {
-  return tableData.map((tableRowData, rowNumber) => (
-    <TableRow key={rowNumber}>
-      {createTableRow({ tableRowData, headingKeys, rowNumber })}
-    </TableRow>
-  ));
-};
-
-const createPageButtons: CreatePageButtons = ({
-  pageSize,
-  totalRecords,
-  setPageNumber,
-}) => {
-  const pageCount = Math.ceil(totalRecords / pageSize);
-  const pageButtons = [];
-  for (let i = 0; i < pageCount; i++) {
-    const pageNumber = i + 1;
-    const label = pageNumber.toString();
-    // TODO: This key needs to be changed obviously! :)
-    pageButtons.push(
-      <Button
-        key={label}
-        label={label}
-        onPress={() => setPageNumber(pageNumber)}
-      ></Button>
-    );
-  }
-  return pageButtons;
-};
 
 // This is the simplest example of a table
 // - NO selection
@@ -173,14 +59,12 @@ export const InMemoryTable = (props: InMemoryTableProps<Object>) => {
   const headingKeys = Object.keys(tableHeaderConfig);
 
   const sortedTableData = sortRows({ sortState, tableData });
-  const rows = createTableRows({ tableData: sortedTableData, headingKeys });
+  const body = createTableBody({ tableData: sortedTableData, headingKeys });
 
   return (
     <Table>
-      <TableHead>
-        <TableRow>{heading}</TableRow>
-      </TableHead>
-      <TableBody>{rows}</TableBody>
+      {heading}
+      {body}
     </Table>
   );
 };
@@ -212,40 +96,26 @@ export const InMemoryPaginatedTable = (
   const sortedTableData = sortRows({ sortState, tableData });
   const [pageNumber, setPageNumber] = useState(1);
 
-  const totalRecords = tableData.length;
   const tablePage = getPage({
     tableData: sortedTableData,
     pageNumber,
     pageSize,
   });
-  const rows = createTableRows({ tableData: tablePage, headingKeys });
-  const pageButtons = createPageButtons({
+  const body = createTableBody({ tableData: tablePage, headingKeys });
+  const paginationControls = createPaginationControls({
+    tableData,
+    pageNumber,
     pageSize,
-    totalRecords,
     setPageNumber,
   });
 
   return (
     <Stack>
       <Table>
-        <TableHead>
-          <TableRow>{heading}</TableRow>
-        </TableHead>
-        <TableBody>{rows}</TableBody>
+        {heading}
+        {body}
       </Table>
-      <Inline>
-        <Button
-          label="<"
-          onPress={() => pageDown({ pageNumber, setPageNumber })}
-        ></Button>
-        {pageButtons}
-        <Button
-          label=">"
-          onPress={() =>
-            pageUp({ pageNumber, pageSize, totalRecords, setPageNumber })
-          }
-        ></Button>
-      </Inline>
+      {paginationControls}
     </Stack>
   );
 };
