+++ src/recipes/tables/types.ts
@@ -22,53 +22,89 @@
 
 import { ReactElement } from "react";
 
-export interface TableHeaderCellConfig {
-    index: boolean,      // Not sure if this is needed
-    sortable: boolean,
-    label: string,
-}
-
-export type TableHeaderConfig<T> = {
-    [Property in keyof T]?: TableHeaderCellConfig
-}
-
-
-// Create React elements for the table heading
-export type CreateTableHeader<T> = (args: { tableHeaderConfig: TableHeaderConfig<T>, sortState: SortState, setSortState: SetSortState}) => ReactElement[];
+/* *****************************************************
+ *
+ * TABLE HEADERS
+ *
+ * *****************************************************/
 
 export type HeadingKey = string;
 
-export type GetHeadingKeys<T> = (args: { tableHeaderConfig: TableHeaderConfig<T> }) => HeadingKey[];
+export type GetHeadingKeys<T> = (args: {
+  tableHeaderConfig: TableHeaderConfig<T>;
+}) => HeadingKey[];
+export interface TableHeaderCellConfig {
+  index: boolean; // Not sure if this is needed
+  sortable: boolean;
+  label: string;
+}
 
-export type CreateTableRow<T> = (args: { tableRowData: T, headingKeys: HeadingKey[], rowNumber: number}) => ReactElement[];
-export type CreateTableRows<T> = (args: { tableData: T[], headingKeys: HeadingKey[]}) => ReactElement[];
-export type CreatePageButtons = (args: { pageSize: number, totalRecords: number, setPageNumber: SetPageNumber}) => ReactElement[];
+export type TableHeaderConfig<T> = {
+  [Property in keyof T]?: TableHeaderCellConfig;
+};
 
-export type SortDirection = 'ASCENDING' | 'DESCENDING';
+// Create React elements for the table heading
+export type CreateTableHeader<T> = (args: {
+  tableHeaderConfig: TableHeaderConfig<T>;
+  sortState: SortState;
+  setSortState: SetSortState;
+}) => ReactElement;
 
-export type GetInitialSortState<T> = (args: { sortAttribute: keyof T, sortDirection: SortDirection}) => SortState;
+/* *****************************************************
+ *
+ * TABLE HEADERS
+ *
+ * *****************************************************/
 
+export type CreateTableRow<T> = (args: {
+  tableRowData: T;
+  headingKeys: HeadingKey[];
+  rowNumber: number;
+}) => ReactElement[];
 
+export type CreateTableBody<T> = (args: {
+  tableData: T[];
+  headingKeys: HeadingKey[];
+}) => ReactElement;
+
+export type CreatePageButtons = (args: {
+  pageSize: number;
+  totalRecords: number;
+  setPageNumber: SetPageNumber;
+}) => ReactElement[];
+
+export type SortDirection = "ASCENDING" | "DESCENDING";
+
+export type GetInitialSortState<T> = (args: {
+  sortAttribute: keyof T;
+  sortDirection: SortDirection;
+}) => SortState;
 
 // Not keen on the name SortState here :/
 export type SortState = {
-    sortDirection: SortDirection,
-    sortAttribute: string,
-}
+  sortDirection: SortDirection;
+  sortAttribute: string;
+};
 
 export type SetSortState = (sortState: SortState) => void;
 
 // Create a predictable sort process...
-export type GetNextSortState = (args: {sortState: SortState, nextSortAttribute: string}) => SortState;
+export type GetNextSortState = (args: {
+  sortState: SortState;
+  nextSortAttribute: string;
+}) => SortState;
 
-export type SortRows<T> = (args: {sortState: SortState, tableData: T[]}) => T[];
+export type SortRows<T> = (args: {
+  sortState: SortState;
+  tableData: T[];
+}) => T[];
 
-
-
-
-export type CreateTableHeaderCell = (args: { headingKey: string, headerConfig: TableHeaderCellConfig, sortState: SortState, setSortState: SetSortState }) => ReactElement;
-
-
+export type CreateTableHeaderCell = (args: {
+  headingKey: string;
+  headerConfig: TableHeaderCellConfig;
+  sortState: SortState;
+  setSortState: SetSortState;
+}) => ReactElement;
 
 // export type RenderHeadCell = (args: { columnNumber: number}) => ReactElement | void; // Allow void to be returned so that data can be used to create some cells
 // export type RenderDataCell = (args: { columnNumber: number, key: string | number}) => ReactElement | void;
@@ -83,7 +119,6 @@ export type CreateTableHeaderCell = (args: { headingKey: string, headerConfig: T
 
 // export type FetchTablePage = (args: { page: number, pageSize: number }) => TableData;
 
-
 // export type DynamicTableProps = {
 //     head?: TableHead; // The data for the heading, probably needs a better name
 //     data?: TableData; // the data for the table body
@@ -97,47 +132,64 @@ export type CreateTableHeaderCell = (args: { headingKey: string, headerConfig: T
 // An "in memory" table is a table that does not fetch data from a service or server. All the data
 // is loaded into memory when created.
 export interface InMemoryTableProps<T> {
-    tableHeaderConfig: TableHeaderConfig<T>,
-    tableData: T[],
-    sortAttribute: keyof T,
-    sortDirection: SortDirection
+  tableHeaderConfig: TableHeaderConfig<T>;
+  tableData: T[];
+  sortAttribute: keyof T;
+  sortDirection: SortDirection;
 }
 
-export type InMemoryTableComponent<T> = (args: InMemoryTableProps<T>) => ReactElement;
+export type InMemoryTableComponent<T> = (
+  args: InMemoryTableProps<T>
+) => ReactElement;
 
 export interface InMemoryPaginatedTableProps<T> extends InMemoryTableProps<T> {
-    pageSize: number,
+  pageSize: number;
 }
-
 
 // export interface InMemoryPaginatedSortableTableProps<T> extends InMemoryPaginatedTableProps<T> {
 //     sortAttribute: [keyof T],
 //     sortDirection: SortDirection,
 // }
 
+export type GetPage<T> = (args: {
+  tableData: T[];
+  pageNumber: number;
+  pageSize: number;
+}) => T[];
 
-
-
-export type GetPage<T> = (args: { tableData: T[], pageNumber: number, pageSize: number }) => T[]
-
-export type PageDown = (args: { pageNumber: number, setPageNumber: SetPageNumber }) => void;
-export type PageUp = (args: { pageNumber: number, pageSize: number, totalRecords: number, setPageNumber: SetPageNumber }) => void;
+export type PageDown = (args: {
+  pageNumber: number;
+  setPageNumber: SetPageNumber;
+}) => void;
+export type PageUp = (args: {
+  pageNumber: number;
+  pageSize: number;
+  totalRecords: number;
+  setPageNumber: SetPageNumber;
+}) => void;
 
 export type SetPageNumber = (args: number) => void;
 
+export type CreatePaginationControls<T> = (args: {
+  tableData: T[];
+  pageNumber: number;
+  pageSize: number;
+  setPageNumber: SetPageNumber;
+}) => ReactElement;
 
-// Should all tables have state? :/ 
+// Should all tables have state? :/
 
-export type DynamicTableState = {
-
-}
+export type DynamicTableState = {};
 
 export const DEFAULT_ACTION = "default";
 
-export type DefaultActionReducer = (args: { state: DynamicTableState, action: DynamicTableAction}) => DynamicTableState;
+export type DefaultActionReducer = (args: {
+  state: DynamicTableState;
+  action: DynamicTableAction;
+}) => DynamicTableState;
 
 export type DefaultDynamicTableAction = {
-    type: typeof DEFAULT_ACTION
-}
+  type: typeof DEFAULT_ACTION;
+};
 
 export type DynamicTableAction = DefaultDynamicTableAction;
